+++ {{ cookiecutter.project_name }}/composition_repository/djangocms_file/disk.py
@@ -1,0 +1,18 @@
+from project_composer.marker import EnabledApplicationMarker
+
+
+class DjangoCmsFileDefinitions(EnabledApplicationMarker):
+    def define(self):
+        return super().define() + [
+            [
+                "djangocms_file",
+                {
+                    "comments": "Djangocms_File",
+                    "natural_foreign": True,
+                    "models": [
+                        "djangocms_file.File",
+                        "djangocms_file.Folder"
+                    ]
+                }
+            ],
+        ]